unify fixture type and group name (#44)
The previous refactoring that eliminated the need to dispatch OSC messages based on the explicit fixture type made the original motivation for tracking a separate group identifier no longer relevant. We can instead use a single string as the fixture group key, using the group name if present and falling back to the fixture type name. This obviates the need for the `:`-prefixed group name wart entirely, allowing us to uniformly treat the first path segment of any OSC address (that doesn't match the few static control groups) as a fixture key and look directly in the patch for it. This also allows removing the lookup map in the patch to retrieve exact `FixtureType` instances.

Huzzah.