streamline subfixtures with animations (#53)
Support animations on subfixtures defined as controls via a new `#[animate_subtarget(...)]` attribute on the `Control` derive macro. This automatically generates target types in the fixture for the named attributes, and implements a conversion trait that allows a simple method call to pass a filtered set of targeted animations to a subfixture.

A bit of refactoring to enable this - `TargetedAnimationValues` now itself owns a fixed-size array, and is passed around by reference. This allows returning a new instance of this from the new `subtarget` convenience method.