Audio monitor on midi wings. (#59)
Make use of the VU meters on the clock wings to display the current audio envelope when we have the monitor switched on. Sexy! Also remember that the MIDI event helper functions exist, and use them. Completely delete support for the AMX headphone buttons because they have too much device-side state management.