Explicit patch options. (#61)
Implement an interface for fixtures to report their patch options, if they have any. Steps towards providing a patching UI. In the interim, provide a help command that provides a listing of all registered fixture types.

Refactors the patcher registry to also include the fixture name and the method that gets the options. This allows us to move the patcher filtering step up into the patch and out of the patch method on the individual fixture.