Explicit group patching. (#63)
Revamp the patch traits, so that creating a group and patching fixtures in it are separate activities. This also revamps the patch file format so that groups are defined exactly once, and then individual patches are specified in each group. Much more explicit and sane.

Splits the options up into two categories - group-level and patch-level, to make it clear which options configure the behavior of the group vs the fixture. This is a slight regression in that there is no longer a separate method/action to patch a fixture in a single group, which we might need for incremental patching. Shouldn't be too hard to refactor this in the future.

We really need tests of the patcher to verify that all of our expected validations are in place... maybe AI will fix it.