Implement support for Behringer CMD DV-1 animation wing. (#67)
The Behringer CMD-fest continues... add MIDI device model and mappings
for initial use of a CMD DV-1 as an animation interface wing. The main
purpose of this is to provide hardware encoders in place of having to
use touch controls to set the main four animation parameters, similar to
how we use the encoders on the APC 40. For the moment, also map an
encoder to the NPeriods parameter (convenient that the encoders have 15
LEDs and we currently allow 0-15 periods), and use the four large square
buttons to perform animation index select.

This required implementing a "Nudge" interface for the animation
parameters. For the moment, this is implemented at the CC level. It
might make more sense to push this down into the underlying animation
control model, since it is potentially useful. The fact that we can give
the encoders any range we want is pretty nice, actually.

Lots of plumbing to wire MIDI emission up to animations.