Tighter strobe clock. (#77)
Adjust the behavior of the strobe clock to internally quantize clock
rates set from a knob, such that the strobe interval is always an
integer number of frames. This provides the most stable strobing
behavior we can get, modulo the lack of synchronization between the
hardware DMX frame clock and the console render clock.

We still use the raw rate when the rate is coming from tap sync, since
these will be slower rates that are less likely to suffer from visible
jitter and staying in sync with the tap is more valuable.

Rescale the global strobe rate control to have quartic behavior,
providing more resolution at the bottom of the range and compressing the
top (which is already highly quantized due to the frame clock).