CLI visualization support (#80)
Add support for fixtures to preview their state into the terminal.
Activated using a CLI flag `--cli-preview`. Support is done via
injection of a preview type, so not all fixtures are required to
implement a preview. Support in place for colors and the cellular
strobes.

Rendering performance is OK but a bit flickery. Should investigate
https://gist.github.com/christianparpart/d8a62cc1ab659194337d73e399004036
for possible synchronization support.