strobe flash distribution (#84)
Revamp the global strobe clock system to "distribute" flashes to groups,
which then manage the flash state internally. This enables the creation
of different distribution modes - so far only "all" and "each group
sequentially" are implemented, and there are no controls for switching
between them (defaults to "all").

- Move responsibility for declaring if a fixture can strobe to the patch
trait, and the fixture declares up-front what length of strobe pulse it
needs.
- Eliminate the StrobeState struct, track if we flashed on the last
update directly in the strobe clock, and delegate things to it.